Graph and Refactor

+++ lib/models/transaction.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 
 class Transaction {
-  final double id;
+  final int id;
   final String name;
   final double amount;
   final DateTime date;
